+++ src/main/scala/calc/ParseTree.scala
@@ -1,7 +1,8 @@
 package calc
 
+import scala.math.BigDecimal
 import Calculator.Op
 
 trait ParseTree
-case class PNum(value: Double) extends ParseTree
+case class PNum(value: BigDecimal) extends ParseTree
 case class POp(value: Op, param0: ParseTree, param1: ParseTree) extends ParseTree
